+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 ### Unreleased
 
+- 2021-10-01 -- Adding function definition to all client functions
+- 2021-09-30 -- Adding ma_script_example.py
+- 2021-09-30 -- Adding retrieving_blocks.mdx
+- 2021-09-30 -- Adding querying_event.mdx
+- 2021-09-30 -- Adding performing_scripts.mdx
+- 2021-09-30 -- Adding retrieving_collections.mdx
+
+### Unreleased
+
 - 2021-09-28 -- Adding ma_block_example.py
 - 2021-09-28 -- Adding ma_events_example.py
 - 2021-09-28 -- Adding ma_collections_example.py
 - 2021-09-28 -- Adding ma_create_account.py
 - 2021-09-28 -- Adding from_seed classmethod AccountKey Class
 - 2021-09-28 -- Adding function definition to some client functions
-- 2021-09-28 -- Adding function definition to some client functions
 - 2021-09-28 -- Adding generating_keys.mdx file
+
+
+
+### Unreleased
+
+- YYYY-MM-DD **BREAKING?** -- description